Relocate Contest*/Training* submission classes into domain packages
The per-domain submission qualifiers and Dagger modules added in #917 sat
in the generic judgels.submission* packages next to the base/interface
infra (SubmissionStore, BaseSubmissionStore, GradingResponsePoller, ...).
Move them so generic code no longer shares a package with domain impls:

- Contest* -> judgels.contest.submission.{programming,bundle}
  (alongside the existing ContestSubmissionResource/ContestItemSubmissionResource)
- Training* + the training-only StatsProcessor (programming & bundle)
  -> new judgels.training.submission.{programming,bundle}

Pure relocation: no @Provides/binding/qualifier-name changes, so the
Dagger graph is identical. Files moved via git mv (history preserved);
imports/FQNs updated across main, test, and integTest source sets
(JudgelsServerComponent, JudgelsServerApplication, contest/training
consumers, tasks, JerahmeelIntegrationTestComponent, StatsIntegrationTests).

Verified: compileJava/compileTestJava/compileIntegTestJava all pass
(Dagger revalidates the merged graph), test suite green (135 tests, 0
failures), no stale old-location references remain.
