+++ judgels-backends/judgels-server-app/src/main/java/judgels/contest/scoreboard/ContestScoreboardUpdaterModule.java
@@ -14,8 +14,8 @@ import judgels.contest.problem.ContestProblemStore;
 import judgels.problem.ProblemService;
 import judgels.profile.ProfileStore;
 import judgels.service.JudgelsScheduler;
-import judgels.submission.ContestSubmissionStore;
-import judgels.submission.bundle.ContestItemSubmissionStore;
+import judgels.contest.submission.programming.ContestSubmissionStore;
+import judgels.contest.submission.bundle.ContestItemSubmissionStore;
 import judgels.submission.bundle.ItemSubmissionStore;
 import judgels.submission.programming.SubmissionStore;
 import org.hibernate.SessionFactory;
