+++ judgels-backends/judgels-server-app/src/main/java/judgels/contest/submission/bundle/ContestItemSubmissionRegrader.java
@@ -1,4 +1,4 @@
-package judgels.submission.programming;
+package judgels.contest.submission.bundle;
 
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
@@ -7,4 +7,4 @@ import java.lang.annotation.Retention;
 
 @Qualifier
 @Retention(RUNTIME)
-public @interface TrainingSubmissionSourceBuilder {}
+public @interface ContestItemSubmissionRegrader {}
