+++ judgels-backends/judgels-server-app/src/main/java/judgels/contest/submission/programming/ContestGradingResponseProcessor.java
@@ -1,4 +1,4 @@
-package judgels.submission;
+package judgels.contest.submission.programming;
 
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
@@ -7,4 +7,4 @@ import java.lang.annotation.Retention;
 
 @Qualifier
 @Retention(RUNTIME)
-public @interface ContestSubmissionStore {}
+public @interface ContestGradingResponseProcessor {}
