+++ judgels-backends/judgels-server-app/src/main/java/judgels/contest/submission/programming/ContestSubmissionModule.java
@@ -1,4 +1,4 @@
-package judgels.submission.programming;
+package judgels.contest.submission.programming;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dagger.Module;
@@ -14,42 +14,52 @@ import judgels.fs.FileSystem;
 import judgels.fs.local.LocalFileSystem;
 import judgels.messaging.MessageClient;
 import judgels.messaging.MessageListener;
-import judgels.persistence.TrainingProgrammingGradingDao;
-import judgels.persistence.TrainingProgrammingSubmissionDao;
+import judgels.persistence.ContestBundleItemSubmissionDao;
+import judgels.persistence.ContestProgrammingGradingDao;
+import judgels.persistence.ContestProgrammingSubmissionDao;
+import judgels.contest.submission.bundle.ContestItemSubmissionStore;
 import judgels.service.JudgelsBaseDataDir;
 import judgels.service.JudgelsScheduler;
-import judgels.stats.StatsConfiguration;
-import judgels.submission.TrainingSubmissionStore;
+import judgels.submission.bundle.BaseItemSubmissionStore;
+import judgels.submission.bundle.ItemSubmissionStore;
+import judgels.submission.programming.BaseSubmissionStore;
+import judgels.submission.programming.GradingResponsePoller;
+import judgels.submission.programming.GradingResponseProcessor;
+import judgels.submission.programming.NoOpSubmissionConsumer;
+import judgels.submission.programming.SubmissionClient;
+import judgels.submission.programming.SubmissionConsumer;
+import judgels.submission.programming.SubmissionDownloader;
+import judgels.submission.programming.SubmissionRegradeProcessor;
+import judgels.submission.programming.SubmissionRegrader;
+import judgels.submission.programming.SubmissionSourceBuilder;
+import judgels.submission.programming.SubmissionStore;
 import org.hibernate.SessionFactory;
 
 @Module
-public class TrainingSubmissionModule {
-    private final StatsConfiguration statsConfig;
+public class ContestSubmissionModule {
     private final Optional<FileSystem> fs;
 
-    public TrainingSubmissionModule(StatsConfiguration statsConfig) {
-        this.statsConfig = statsConfig;
+    public ContestSubmissionModule() {
         this.fs = Optional.empty();
     }
 
-    public TrainingSubmissionModule(StatsConfiguration statsConfig, FileSystem fs) {
-        this.statsConfig = statsConfig;
+    public ContestSubmissionModule(FileSystem fs) {
         this.fs = Optional.of(fs);
     }
 
     @Provides
     @Singleton
-    @TrainingSubmissionFs
+    @ContestSubmissionFs
     FileSystem submissionFs(@JudgelsBaseDataDir Path baseDataDir) {
         return fs.orElse(new LocalFileSystem(baseDataDir.resolve("submissions")));
     }
 
     @Provides
     @Singleton
-    @TrainingSubmissionStore
-    static SubmissionStore trainingSubmissionStore(
-            TrainingProgrammingSubmissionDao submissionDao,
-            TrainingProgrammingGradingDao gradingDao,
+    @ContestSubmissionStore
+    static SubmissionStore submissionStore(
+            ContestProgrammingSubmissionDao submissionDao,
+            ContestProgrammingGradingDao gradingDao,
             ObjectMapper mapper) {
 
         return new BaseSubmissionStore<>(submissionDao, gradingDao, mapper);
@@ -57,17 +67,24 @@ public class TrainingSubmissionModule {
 
     @Provides
     @Singleton
-    @TrainingSubmissionSourceBuilder
-    static SubmissionSourceBuilder submissionSourceBuilder(@TrainingSubmissionFs FileSystem submissionFs) {
+    @ContestItemSubmissionStore
+    static ItemSubmissionStore itemSubmissionStore(ContestBundleItemSubmissionDao submissionDao) {
+        return new BaseItemSubmissionStore<>(submissionDao);
+    }
+
+    @Provides
+    @Singleton
+    @ContestSubmissionSourceBuilder
+    static SubmissionSourceBuilder submissionSourceBuilder(@ContestSubmissionFs FileSystem submissionFs) {
         return new SubmissionSourceBuilder(submissionFs);
     }
 
     @Provides
     @Singleton
-    @TrainingSubmissionClient
+    @ContestSubmissionClient
     static SubmissionClient submissionClient(
             SessionFactory sessionFactory,
-            @TrainingSubmissionStore SubmissionStore submissionStore,
+            @ContestSubmissionStore SubmissionStore submissionStore,
             @Named("gradingRequestQueueName") String gradingRequestQueueName,
             @Named("gradingResponseQueueName") String gradingResponseQueueName,
             MessageClient messageClient,
@@ -84,14 +101,21 @@ public class TrainingSubmissionModule {
 
     @Provides
     @Singleton
-    @TrainingSubmissionRegrader
+    static SubmissionDownloader submissionDownloader(
+            @ContestSubmissionSourceBuilder SubmissionSourceBuilder submissionSourceBuilder) {
+        return new SubmissionDownloader(submissionSourceBuilder);
+    }
+
+    @Provides
+    @Singleton
+    @ContestSubmissionRegrader
     static SubmissionRegrader submissionRegrader(
             JudgelsScheduler scheduler,
-            @TrainingSubmissionStore SubmissionStore submissionStore,
-            @TrainingSubmissionSourceBuilder SubmissionSourceBuilder submissionSourceBuilder,
-            @TrainingSubmissionClient SubmissionClient submissionClient) {
+            @ContestSubmissionStore SubmissionStore submissionStore,
+            @ContestSubmissionSourceBuilder SubmissionSourceBuilder submissionSourceBuilder,
+            @ContestSubmissionClient SubmissionClient submissionClient) {
 
-        ExecutorService executorService = scheduler.createExecutorService("training-submission-regrade-processor-%d", 5);
+        ExecutorService executorService = scheduler.createExecutorService("contest-submission-regrade-processor-%d", 5);
         return new SubmissionRegrader(
                 submissionStore,
                 executorService,
@@ -100,14 +124,14 @@ public class TrainingSubmissionModule {
 
     @Provides
     @Singleton
-    @TrainingGradingResponsePoller
+    @ContestGradingResponsePoller
     static GradingResponsePoller gradingResponsePoller(
             JudgelsScheduler scheduler,
             MessageListener messageListener,
             @Named("gradingResponseQueueName") String gradingResponseQueueName,
-            @TrainingGradingResponseProcessor GradingResponseProcessor processor) {
+            @ContestGradingResponseProcessor GradingResponseProcessor processor) {
 
-        ExecutorService executorService = scheduler.createExecutorService("training-grading-response-processor-%d", 10);
+        ExecutorService executorService = scheduler.createExecutorService("contest-grading-response-processor-%d", 10);
         return new GradingResponsePoller(
                 messageListener,
                 gradingResponseQueueName,
@@ -117,12 +141,11 @@ public class TrainingSubmissionModule {
 
     @Provides
     @Singleton
-    @TrainingGradingResponseProcessor
-    GradingResponseProcessor gradingResponseProcessor(
+    @ContestGradingResponseProcessor
+    static GradingResponseProcessor gradingResponseProcessor(
             UnitOfWorkAwareProxyFactory unitOfWorkAwareProxyFactory,
             ObjectMapper mapper,
-            @TrainingSubmissionStore SubmissionStore submissionStore,
-            StatsProcessor statsProcessor) {
+            @ContestSubmissionStore SubmissionStore submissionStore) {
 
         return unitOfWorkAwareProxyFactory.create(
                 GradingResponseProcessor.class,
@@ -133,6 +156,6 @@ public class TrainingSubmissionModule {
                 new Object[] {
                         mapper,
                         submissionStore,
-                        statsConfig.getEnabled() ? statsProcessor : new NoOpSubmissionConsumer()});
+                        new NoOpSubmissionConsumer()});
     }
 }
