+++ judgels-backends/judgels-server-app/src/main/java/judgels/contest/submission/programming/ContestSubmissionResource.java
@@ -66,10 +66,6 @@ import judgels.problem.ProblemUtils;
 import judgels.profile.ProfileStore;
 import judgels.service.actor.ActorChecker;
 import judgels.service.api.actor.AuthHeader;
-import judgels.submission.ContestSubmissionStore;
-import judgels.submission.programming.ContestSubmissionClient;
-import judgels.submission.programming.ContestSubmissionRegrader;
-import judgels.submission.programming.ContestSubmissionSourceBuilder;
 import judgels.submission.programming.SubmissionClient;
 import judgels.submission.programming.SubmissionDownloader;
 import judgels.submission.programming.SubmissionRegrader;
