+++ judgels-backends/judgels-server-app/src/main/java/judgels/tasks/JerahmeelTaskModule.java
@@ -4,9 +4,9 @@ import dagger.Module;
 import dagger.Provides;
 import io.dropwizard.hibernate.UnitOfWorkAwareProxyFactory;
 import jakarta.inject.Singleton;
-import judgels.submission.ContestSubmissionStore;
-import judgels.submission.TrainingSubmissionStore;
-import judgels.submission.programming.StatsProcessor;
+import judgels.contest.submission.programming.ContestSubmissionStore;
+import judgels.training.submission.programming.TrainingSubmissionStore;
+import judgels.training.submission.programming.StatsProcessor;
 import judgels.submission.programming.SubmissionStore;
 
 @Module
