+++ judgels-backends/judgels-server-app/src/main/java/judgels/training/submission/bundle/TrainingItemSubmissionRegrader.java
@@ -1,4 +1,4 @@
-package judgels.submission;
+package judgels.training.submission.bundle;
 
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
@@ -7,4 +7,4 @@ import java.lang.annotation.Retention;
 
 @Qualifier
 @Retention(RUNTIME)
-public @interface TrainingSubmissionStore {}
+public @interface TrainingItemSubmissionRegrader {}
