+++ judgels-backends/judgels-server-app/src/main/java/judgels/training/submission/programming/TrainingSubmissionClient.java
@@ -1,4 +1,4 @@
-package judgels.submission.programming;
+package judgels.training.submission.programming;
 
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
@@ -7,4 +7,4 @@ import java.lang.annotation.Retention;
 
 @Qualifier
 @Retention(RUNTIME)
-public @interface TrainingSubmissionFs {}
+public @interface TrainingSubmissionClient {}
